Changed reading types to decimal.

+++ AirPollutionWebApi/Models/Reading.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public int TimeStamp { get; set; }
-        public int Co { get; set; }
-        public int No { get; set; }
-        public int So { get; set; }
+        public decimal Co { get; set; }
+        public decimal No { get; set; }
+        public decimal So { get; set; }
     }
 }
